+++ iaso/tests/api/validation_workflows/test_views/common.py
@@ -1,0 +1,26 @@
+from iaso.test import APITestCase
+
+
+class BaseValidationWorkflowAPITestCase(APITestCase):
+    def assertValidValidationWorkflowListData(self, list_data, expected_length, paginated=True):
+        results_key = "results"
+        self.assertValidListData(
+            list_data=list_data, results_key=results_key, expected_length=expected_length, paginated=False
+        )
+
+        for data in list_data[results_key]:
+            self.assertIn("slug", data)
+            self.assertIn("name", data)
+            self.assertIn("formCount", data)
+            self.assertIn("createdBy", data)
+            self.assertIn("updatedBy", data)
+            self.assertIn("createdAt", data)
+            self.assertIn("updatedAt", data)
+
+    def assertValidValidationWorkflowDropdownListData(self, list_data, expected_length):
+        self.assertValidListData(
+            list_data=list_data, results_key=None, expected_length=expected_length, paginated=False
+        )
+        for data in list_data:
+            self.assertIn("label", data)
+            self.assertIn("value", data)
